fix(spa): address PR review findings
- #5: Update electron.d.ts with 9 new IPC method declarations
- #6: Remove unused title prop from BrowserToolbar and callers
- #8: Extract useBrowserViewResize hook to eliminate duplication
