+++ spa/src/types/electron.d.ts
@@ -61,6 +61,15 @@ interface Window {
 
     // Browser View
     resizeBrowserView: (paneId: string, bounds: ElectronBounds) => Promise<void>
+    browserViewGoBack: (paneId: string) => Promise<void>
+    browserViewGoForward: (paneId: string) => Promise<void>
+    browserViewReload: (paneId: string) => Promise<void>
+    browserViewStop: (paneId: string) => Promise<void>
+    destroyBrowserView: (paneId: string) => Promise<void>
+    browserViewOpenMiniWindow: (url: string) => Promise<void>
+    browserViewMoveToTab: (paneId: string) => Promise<void>
+    onBrowserViewStateUpdate: (callback: (paneId: string, state: { url: string; title: string; canGoBack: boolean; canGoForward: boolean; isLoading: boolean }) => void) => () => void
+    onBrowserViewOpenInTab: (callback: (url: string) => void) => () => void
 
     // Memory Monitor
     getProcessMetrics: () => Promise<ElectronTabMetrics[]>
